Refactorización de código en componente SelectInput
